python/requirements.txt
google-generativeai
python-dotenv
pillow
fastapi
uvicorn
cloudinary
python-dotenv